getter.jsをtsに変換。かつ、書籍情報取得処理を追加した。
